@@ -13,18 +13,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.3.12",
+    "@angular/animations": "^14.3.0",
     "@angular/cdk": "^13.3.9",
-    "@angular/common": "^13.4.0",
-    "@angular/core": "^13.4.0",
-    "@angular/forms": "^13.4.0",
+    "@angular/common": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
     "@angular/google-maps": "^13.3.9",
     "@angular/material": "^13.3.9",
-    "@angular/platform-browser": "^13.4.0",
-    "@angular/platform-browser-dynamic": "^13.4.0",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
     "@angular/pwa": "^12.1.0",
-    "@angular/router": "^13.4.0",
-    "@angular/service-worker": "^13.4.0",
+    "@angular/router": "^14.3.0",
+    "@angular/service-worker": "^14.3.0",
     "@capacitor/android": "4.0.1",
     "@capacitor/app": "4.0.1",
     "@capacitor/camera": "4.0.1",
@@ -87,15 +87,15 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.11",
+    "@angular-devkit/build-angular": "^14.2.13",
     "@angular-eslint/builder": "^13.5.0",
     "@angular-eslint/eslint-plugin": "~12.0.0",
     "@angular-eslint/eslint-plugin-template": "~12.0.0",
     "@angular-eslint/template-parser": "~12.0.0",
-    "@angular/cli": "~13.3.11",
-    "@angular/compiler": "~13.4.0",
-    "@angular/compiler-cli": "~13.4.0",
-    "@angular/language-service": "~13.4.0",
+    "@angular/cli": "^14.2.13",
+    "@angular/compiler": "^14.3.0",
+    "@angular/compiler-cli": "^14.3.0",
+    "@angular/language-service": "^14.3.0",
     "@capacitor/cli": "4.0.1",
     "@ionic/angular-toolkit": "^4.0.0",
     "@ionic/lab": "^3.2.10",
